+++ src/cli/commands/commands/generate/validate-global-config.ts
@@ -6,9 +6,23 @@ import {
   isBlankString,
 } from '../../util';
 
+export interface ValidateGlobalConfigResultError {
+  readonly isValid: false;
+  readonly errors: readonly string[];
+}
+
+export interface ValidateGlobalConfigResultSuccess {
+  readonly isValid: true;
+  readonly globalConfig: GlobalConfig;
+}
+
+export type ValidateGlobalConfigResult =
+  | ValidateGlobalConfigResultError
+  | ValidateGlobalConfigResultSuccess;
+
 export function validateGlobalConfig(
   globalConfig: Partial<GlobalConfig>,
-): readonly string[] {
+): ValidateGlobalConfigResult {
   const errors: string[] = [];
 
   for (const optionName of GLOBAL_CONFIG_OPTION_NAMES) {
@@ -22,5 +36,13 @@ export function validateGlobalConfig(
     }
   }
 
-  return errors;
+  return errors.length === 0
+    ? {
+        isValid: true,
+        globalConfig: globalConfig as GlobalConfig,
+      }
+    : {
+        isValid: false,
+        errors,
+      };
 }