+++ src/cli/commands/types/generator-inputs.ts
@@ -1,7 +1,0 @@
-import { ProjectDataAny } from './project-kind';
-
-export interface GeneratorInputs {
-  readonly output: string;
-  readonly projectName: string;
-  readonly projectData: ProjectDataAny;
-}