+++ src/cli/commands/types/index.ts
@@ -1,5 +1,1 @@
 export * from './global-config';
-export * from './generator-inputs';
-export * from './project-kind';
-export * from './template-app';
-export * from './template-lib';